Consolidate basic energy inventory
Instead of each harvester checking all the spawns and extensions each
tick, the main loop checks each room every few ticks, and the creeps
check those cached values, only doing the extra computation if the
values indicate an energy deficit. This also helped to ease the cpu on
the can_withdraw2 function which is replacing the can_withdraw function.