Enable conditional renewing of restockers
The renew function will only renew the creep now if the creeps parts are
the same as the parts designated for its role in the room_strategy. This
allows old creeps to die and new creeps replace them if the
room_strategy for them changes. This automatic renewal is only enabled
for role_restockers at the moment.